Fixed texturing for bar style and make tiling resolution-independent. Always uses virtual resolution of 1920 x 1080, option to change by user could be implemented.

+++ mg_visual music/NoteStyle_Bar.cs
@@ -122,7 +122,7 @@ namespace Visual_Music
 				Color color;
 				Texture2D texture;
 
-				getMaterial(trackProps, (int)(noteStart.X - songPosP), (int)(noteEnd.X - songPosP), out color, out texture);
+				getMaterial(trackProps, noteStart.X - songPosP, noteEnd.X - songPosP, out color, out texture);
 				//fx.Parameters["Color"].SetValue(color.ToVector4());
 				instanceVerts[n].color = color;
 				fx.Parameters["Texture"].SetValue(texture);
